Removing `ContentPackInstallationHook`, extracting `ContentPackEntityResolver`, adjusting tests.
