+++ graylog2-server/src/main/java/org/graylog2/contentpacks/ContentPacksModule.java
@@ -74,7 +74,5 @@ public class ContentPacksModule extends PluginModule {
         registerJacksonSubtype(MessageListEntity.class);
         registerJacksonSubtype(PivotEntity.class);
         registerJacksonSubtype(EventListEntity.class);
-
-        addContentPackInstallationHook(ShareEntitiesHook.class);
     }
 }
