+++ graylog2-server/src/main/java/org/graylog2/plugin/PluginModule.java
@@ -51,7 +51,6 @@ import org.graylog.security.shares.SyncedEntitiesResolver;
 import org.graylog2.audit.AuditEventType;
 import org.graylog2.audit.PluginAuditEventTypes;
 import org.graylog2.audit.formatter.AuditEventFormatter;
-import org.graylog2.contentpacks.ContentPackInstallationHook;
 import org.graylog2.contentpacks.constraints.ConstraintChecker;
 import org.graylog2.contentpacks.facades.EntityWithExcerptFacade;
 import org.graylog2.contentpacks.model.ModelType;
@@ -501,13 +500,5 @@ public abstract class PluginModule extends Graylog2Module {
 
             dbEntitiesBinder().addBinding().toInstance(entitiesClass);
         }
-    }
-
-    protected void addContentPackInstallationHook(Class<? extends ContentPackInstallationHook> contentPackInstallationHook) {
-        contentPackInstallationHookBinder().addBinding().to(contentPackInstallationHook);
-    }
-
-    protected Multibinder<ContentPackInstallationHook> contentPackInstallationHookBinder() {
-        return Multibinder.newSetBinder(binder(), ContentPackInstallationHook.class);
     }
 }
